STT modularized
